SpringBoot and SpringWeb features need Spring
